fix(conductor): preflight 500 + dispatcher ::bigint + logging
FIX A — Preflight 500: sync_type sbagliato
- preflightPending cercava sync_type='sync-products' ma la tabella
  shared.sync_sessions ha CHECK sync_type='products'. Con sync_sessions
  vuota il return era precoce (changes:[]) ma con dati futuri avrebbe
  filtrato 0 righe per errore. Corretto in 'products'.

FIX B — Preflight 500: DISTINCT ON senza ORDER BY
- SELECT DISTINCT ON (p.product_id) senza ORDER BY p.product_id →
  PostgreSQL lancia "SELECT DISTINCT ON expressions must match initial
  ORDER BY expressions" → catch → 500. Aggiunto ORDER BY p.product_id.
  Questo era la causa root del 500 se il pending aveva confirmed_at
  valorizzato e sync_sessions non era vuota.

FIX D — Preflight: catch silenzioso
- Il catch nella route non loggava l'errore: impossible diagnosticare.
  Aggiunto logger.error con pendingId e messaggio.

FIX E — dispatcher.ts: WHERE task_id = $1::bigint
- I 2 callback del recoverOrphans (resumeFromErpSaveDone, reEnqueueTask)
  usavano WHERE task_id = $1 senza ::bigint → potenziale "operator is
  not unique: unknown - unknown" durante recovery on startup.

+++ archibald-web-app/backend/src/conductor/dispatcher.ts
@@ -47,7 +47,7 @@ export class Conductor extends EventEmitter {
         await this.deps.pool.query(
           `UPDATE system.agent_operation_queue
            SET status = 'enqueued', started_at = NULL, heartbeat_at = NULL
-           WHERE task_id = $1`,
+           WHERE task_id = $1::bigint`,
           [task.taskId.toString()],
         );
       },
@@ -61,7 +61,7 @@ export class Conductor extends EventEmitter {
                erp_order_id = NULL,
                started_at = NULL,
                heartbeat_at = NULL
-           WHERE task_id = $1`,
+           WHERE task_id = $1::bigint`,
           [task.taskId.toString()],
         );
       },
